LayoutTests/fast/events/spatial-navigation/snav-iframe-recursive-offset-parent-expected.txt

Link

PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "3"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "2"
PASS document.defaultView.frames[0].document.activeElement.getAttribute("id") is "1"
PASS document.activeElement.getAttribute("id") is "start"
